remove_dir_recursively(): check on actual dir

+++ src/pkgbuild.rs
@@ -499,7 +499,7 @@ fn remove_dir_recursively<P: AsRef<Path>>(dir: P) -> Result<(), std::io::Error>
     for entry in read_dir(dir)? {
         let entry = entry?;
         let path = entry.path();
-        if path.is_dir() {
+        if !path.is_symlink() && path.is_dir() {
             let er = remove_dir_recursively(&path);
             match remove_dir(&path) {
                 Ok(_) => (),
